Split up search serializer tests
